requirements.txt
ManimPango==0.6.0
Markdown==3.7
MarkupSafe==3.0.2
absl-py==2.1.0
astunparse==1.6.3
av==13.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloup==3.0.6
colorama==0.4.6
decorator==5.2.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.6.0
glcontext==3.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
isosurfaces==0.1.2
jinja2==3.1.6
keras==3.9.0
libclang==18.1.1
manim==0.19.0
mapbox-earcut==1.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.5.1
moderngl-window==3.1.1
moderngl==5.12.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
opt-einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
pip==21.2.3
protobuf==5.29.3
pycairo==1.27.0
pydub==0.25.1
pyglet==2.1.3
pyglm==2.8.0
pygments==2.19.1
requests==2.32.3
rich==13.9.4
scipy==1.15.2
screeninfo==0.8.1
setuptools==57.4.0
six==1.17.0
skia-pathops==0.8.0.post2
soupsieve==2.6
srt==3.5.3
svgelements==1.9.6
sympy==1.13.1
tensorboard-data-server==0.7.2
tensorboard==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.19.0
termcolor==2.5.0
torch==2.6.0
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
